requirements.txt
# Function dependencies, for example:
# package>=version
psycopg2-binary